Allow VFS module to be defined as external lib (#33)
* Fix level_string entry order

The level_string array order need to match the
CHIMERA_LOG level

Signed-off-by: Alain Renaud <alain.renaud@quantum.com>

* Allow loading external VFS modules

When a module is added to the server config like this
 "server": {
    "vfs": {
       "custom": {
           "path": "/usr/local/lib/libchimera_vfs_custom.so",
           "config": "/usr/local/etc/custom.json"
       }
    }
 },

we need to be able to load that exernal module in the
system.

Signed-off-by: Alain Renaud <alain.renaud@quantum.com>

* set chimera.json path based on INSTALLDIR

Set the default config path based on the CMAKE_INSTALL_PREFIX

Signed-off-by: Alain Renaud <alain.renaud@quantum.com>

* Don't attempt load module that are pre-loaded

Chimera has a list of modules that are pre-loaded at
compile time. When adding the module to the server
VFS section do not attempt to load them again if
they are already load. Also allow VFS entry that
do not provide a path if the module is already loaded.

Signed-off-by: Alain Renaud <alain.renaud@quantum.com>

---------

Signed-off-by: Alain Renaud <alain.renaud@quantum.com>